nts/pyproject.toml
[project]
name = "novus-time-server"
version = "1.0.0"
description = "Novus Time Server Configuration Tool"
requires-python = ">=3.8"
dependencies = ["click>=8.1", "pyserial>=3.5"]

[project.scripts]
nts = "nts:cli"

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"

[tool.flit.module]
name = "nts"
